added the VideoCardModal
This card will be charged with displaying a video within the modal. As of right now, it is just displaying the image, and this is meant to be just the template of the modal for now. Changes will happen so that the modal will play and render text correctly as desired.